requirements.txt
smooth_operator_agent_tools>=1.0.105
openai>=1.0.0
python-dotenv>=1.0.0 